requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.10
diffusers==0.31.0
duckduckgo_search==6.3.7
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
networkx==3.4.2
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
primp==0.8.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.3
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.0
transformers==4.46.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wheel==0.45.0
wrapt==1.17.0
yarl==1.18.0
zipp==3.21.0
